rag_system/api_gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
qdrant-client==1.12.0
google-genai==0.3.0
httpx==0.25.2
python-dotenv==1.0.0
